added new apis

+++ server/app/helpers/cleanup-files.js
@@ -1,6 +1,7 @@
 import { deleteFile } from "../utils/file.js";
 
 export const cleanupFiles = async (filePaths) => {
+  if (!Array.isArray(filePaths) || !filePaths?.length) return;
   for (const filePath of filePaths) {
     try {
       await deleteFile(filePath);
